Add debounce utility and fetch user ID function; update item listing API ; Add Feat to fetch item from list and enhance styling

+++ app/server/listItem.action.ts
@@ -1,44 +1,53 @@
 import { FoodItemType } from "@/types/FoodItem";
 import { fetchUserEmail } from "@/utils/fetchUserEmail";
+import { fetchUserId } from "@/utils/fetchUserId";
 import prisma from "@/utils/prisma";
+import { AllCategory } from "@prisma/client";
 
-const fetchUserId = async () => {
-    const userEmail = await fetchUserEmail();
-    if (!userEmail)
-        throw new Error(`User not found`);
+const isExistedItem = async (itemId: string, userId: string) => {
+    return await prisma.foodItemList.findUnique({
+        where: {
+            user_item_id: itemId,
+            user_id: userId!,
+        }
+    });
+}
 
-    const userId = await prisma.user.findFirst({ where: { email: userEmail } }).then(user => user?.id);
-
-    if (!userId) {
-        throw new Error("User ID not found");
-    }
-
-    return userId;
+const isDuplicateItem = async (foodName: string, userID: string) => {
+    return await prisma.foodItemList.count({
+        where: {
+            user_id: userID,
+            name: foodName,
+        }
+    });
 }
 
 export const addItemToList = async (foodItem: FoodItemType, createdItemId: string) => {
     try {
         const userId = await fetchUserId();
 
-        if (!createdItemId)
-            throw new Error(`Item ID not Found`);
+        if (!userId || !createdItemId)
+            throw new Error(`User or Item ID not Found`);
 
-        await prisma.foodItemList.create({
-            data: {
-                name: foodItem.name,
-                currentWeight: foodItem.currentWeight,
-                calories: foodItem.calories,
-                protein: foodItem.protein,
-                carbs: foodItem.carbs,
-                fat: foodItem.fat,
-                sugar: foodItem.sugar,
-                amountPer: foodItem.amountPer,
-                category: foodItem.category,
-                listed: foodItem.listed,
-                user_id: userId!,
-                user_item_id: createdItemId,
-            }
-        })
+        const duplicateItem = await isDuplicateItem(foodItem.name, userId);
+
+        if (!duplicateItem) {
+            await prisma.foodItemList.create({
+                data: {
+                    name: foodItem.name,
+                    calories: foodItem.calories,
+                    protein: foodItem.protein,
+                    carbs: foodItem.carbs,
+                    fat: foodItem.fat,
+                    sugar: foodItem.sugar,
+                    amountPer: foodItem.amountPer,
+                    category: foodItem.category,
+                    listed: foodItem.listed,
+                    user_id: userId!,
+                    user_item_id: createdItemId,
+                }
+            })
+        }
 
     } catch (error: any) {
         throw new Error(error.message || `Failed to add item to list`);
@@ -49,23 +58,28 @@ export const updateItemFromList = async (foodItem: FoodItemType) => {
     try {
         const userId = await fetchUserId();
 
-        await prisma.foodItemList.update({
-            where: {
-                user_item_id: foodItem.id,
-                user_id: userId!,
-            },
-            data: {
-                name: foodItem.name,
-                currentWeight: foodItem.currentWeight,
-                calories: foodItem.calories,
-                protein: foodItem.protein,
-                carbs: foodItem.carbs,
-                fat: foodItem.fat,
-                sugar: foodItem.sugar,
-                amountPer: foodItem.amountPer,
-                category: foodItem.category,
-            }
-        })
+        if (!userId || !foodItem.id) throw new Error(`User or Item not found`);
+
+        const item = await isExistedItem(foodItem.id, userId);
+
+        if (item) {
+            await prisma.foodItemList.update({
+                where: {
+                    user_item_id: foodItem.id,
+                    user_id: userId!,
+                },
+                data: {
+                    name: foodItem.name,
+                    calories: foodItem.calories,
+                    protein: foodItem.protein,
+                    carbs: foodItem.carbs,
+                    fat: foodItem.fat,
+                    sugar: foodItem.sugar,
+                    amountPer: foodItem.amountPer,
+                    category: foodItem.category,
+                }
+            })
+        }
 
     } catch (error: any) {
         throw new Error(error.message || `Failed to update item in list`);
@@ -76,12 +90,18 @@ export const deleteItemFromList = async (itemId: string) => {
     try {
         const userId = await fetchUserId();
 
-        await prisma.foodItemList.delete({
-            where: {
-                user_item_id: itemId,
-                user_id: userId!,
-            }
-        })
+        if (!userId || !itemId) throw new Error(`User not found`);
+
+        const item = await isExistedItem(itemId, userId);
+
+        if (item) {
+            await prisma.foodItemList.delete({
+                where: {
+                    user_item_id: itemId,
+                    user_id: userId,
+                }
+            });
+        }
 
     } catch (error: any) {
         throw new Error(error.message || `Failed to delete item from list`);
@@ -92,9 +112,11 @@ export const fetchItemListOfUser = async () => {
     try {
         const userId = await fetchUserId();
 
+        if (!userId) throw new Error(`User not found`);
+
         return await prisma.foodItemList.findMany({
             where: {
-                user_id: userId!,
+                user_id: userId,
             }
         })
